// HrPro-Recruit-main/src/types/jobs.types.ts
export interface JobType {
  id: string;
  title: string;
  description: string;
  type: EmployementType;
  category: string;
  workMode: WorkMode;
  currency: Currency;
  application?: string | null;
  skills: string[];
  expired: boolean;
  hasExpiryDate: boolean;
  expiryDate?: Date | null;
  hasSalaryRange: boolean;
  minSalary?: number | null;
  maxSalary?: number | null;
  hasExperiencerange: boolean;
  minExperience?: number | null;
  maxExperience?: number | null;
  isVerifiedJob: boolean;
  deleted: boolean;
  deletedAt?: Date | null;
  postedAt: Date;
  updatedAt: Date;
  responsibilities: string[];
  customQuestions?: unknown; // Or specify structure if known
  rejectionMessage?: string | null;
  acceptanceMessage?: string | null;
  company: CompanyType;
}

export interface CompanyType {
  id: string;
  companyName: string;
  companyLogo?: string | null;
  companyEmail: string;
  companyBio: string;
  foundedYear: string;
  numberOfEmployees: string;
  CEOName: string;
  companyType: string;
  city: string;
  country: string;
  website?: string | null;
  linkedinLink?: string | null;
  twitterLink?: string| null;
}